미들웨어 동작방식 변경, 전체적인 구조 변경
- 전체적인 구조 정리 (lib 디렉터리 하위)
    - Authorization: oauth-server 역할: token introspect & validation, scope validation
    - Constant: 내부 OAuth2 구현을 위한 상수들
    - Grant: oauth-client 역할: grant 요청
    - Symfony: Symfony(Silex) 미들웨어 제공
- `Introspector` 네임스페이스 삭제: JWT 파싱과 관련된 부분은 `lcobucci/jwt` 패키지에 의존하도록 함
- 미들웨어를 세 종류로 제공
    - `introspect()`: access_token이 존재할 경우 토큰 파싱 및 유효성 검사
    - `loginRequired()`:
    - `checkScope()`: 이전과 동일
- `RidiTokenInfo`: 파싱된 토큰 정보의 기능 확장 (3가지 메서드 포함)
    - `isExpired()`: 토큰이 만료되었는지를 반환
    - `isValid()`: 토큰 payload의 필수 값들 검증
    - `hasScope()`: 특정 scope를 가지고 있는지를 검사
- 클래스 네이밍 수정
- 테스트케이스 수정

---

- 처음에는 [`league/oauth2-server`의 `AuthorizationValidatorInterface`](https://github.com/thephpleague/oauth2-server/blob/master/src/AuthorizationValidators/AuthorizationValidatorInterface.php)를 구현해 introspector를 대체하려고 했으나, 다음의 이유로 `league/oauth2-server`에서 사용하는 `lcobucci/jwt`를 활용해 직접 구현하는 것으로 변경함
    - `Request` 객체의 변환이 필요하고 이를 위해 불필요한 부가 작업, 디펜던시가 포함되게 됨
    - `AuthorizationValidatorInterface`는 이 라이브러리 사용자가 직접 다루는 인터페이스이므로, 확장해도 별 다른 효용성이 없음
