+++ composer.json
@@ -5,12 +5,12 @@
   "type": "library",
   "license": "MIT",
   "require": {
-    "firebase/php-jwt": "^4.0",
     "symfony/http-foundation": "^2.1",
-    "symfony/http-kernel": "~3.0"
+    "symfony/http-kernel": "~3.0",
+    "lcobucci/jwt": "^3.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7"
+    "phpunit/phpunit": "^6"
   },
   "autoload": {
     "psr-4": {
