+++ lib/Authorization/Exception/AuthorizationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ridibooks\OAuth2Resource\Authorization\Exception;
+
+
+class AuthorizationException extends \Exception
+{
+
+}
