+++ lib/Authorization/Exception/InvalidJwtException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Ridibooks\OAuth2Resource\Authorization\Exception;
+
+class InvalidJwtException extends AuthorizationException
+{
+}
