--- lib/Client/Constant/OAuth2GrantType.php
@@ -1,11 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Client\Constant;
-
-class OAuth2GrantType
-{
-    const AUTHORIZATION_CODE = 'authorization_code';
-    const IMPLICIT = 'implicit';
-    const RESOURCE_OWNER_PASSWORD_CREDENTIALS = 'resource_owner_password_credentials';
-    const CLIENT_CREDENTIALS = 'client_credentials';
-    const REFRESH_TOKEN = 'refresh_token';
-}
