--- lib/Client/Exception/InvalidResponseException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Client\Exception;
-
-use Ridibooks\OAuth2Resource\Common\Exception\BaseException;
-
-class InvalidResponseException extends BaseException
-{
-}
