--- lib/Client/Exception/NotSupportedGrantException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Client\Exception;
-
-use Ridibooks\OAuth2Resource\Common\Exception\BaseException;
-
-class NotSupportedGrantException extends BaseException
-{
-}
