--- lib/Common/Constant/TokenType.php
@@ -1,7 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Common\Constant;
-
-class TokenType
-{
-    const BEARER = 'BEARER';
-}
