--- lib/Common/Exception/BaseException.php
@@ -1,6 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Common\Exception;
-
-abstract class BaseException extends \Exception
-{
-}
