+++ lib/Constant/GrantTypeConstant.php
@@ -1,0 +1,11 @@
+<?php
+namespace Ridibooks\OAuth2Resource\Constant;
+
+class GrantTypeConstant
+{
+    const AUTHORIZATION_CODE = 'authorization_code';
+    const IMPLICIT = 'implicit';
+    const RESOURCE_OWNER_PASSWORD_CREDENTIALS = 'resource_owner_password_credentials';
+    const CLIENT_CREDENTIALS = 'client_credentials';
+    const REFRESH_TOKEN = 'refresh_token';
+}
