+++ lib/Constant/ScopeConstant.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ridibooks\OAuth2Resource\Constant;
+
+
+class ScopeConstant
+{
+    const ALL = 'all';
+}
