+++ lib/Grant/Exception/GrantException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Ridibooks\OAuth2Resource\Grant\Exception;
+
+
+class GrantException extends \Exception
+{
+
+}
