+++ lib/Grant/Exception/InvalidResponseException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Ridibooks\OAuth2Resource\Grant\Exception;
+
+
+class InvalidResponseException extends GrantException
+{
+}
