+++ lib/Grant/Exception/NotSupportedGrantException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Ridibooks\OAuth2Resource\Grant\Exception;
+
+
+class NotSupportedGrantException extends GrantException
+{
+}
