--- lib/Introspector/Exception/ExpireTokenException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Introspector\Exception;
-
-use Ridibooks\OAuth2Resource\Common\Exception\BaseException;
-
-class ExpireTokenException extends BaseException
-{
-}
