--- lib/Introspector/Exception/InvalidJwtSignatureException.php
@@ -1,8 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Introspector\Exception;
-
-use Ridibooks\OAuth2Resource\Common\Exception\BaseException;
-
-class InvalidJwtSignatureException extends BaseException
-{
-}
