--- lib/Resource/Constant/Scope.php
@@ -1,7 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Resource\Constant;
-
-class Scope
-{
-    const ALL = 'all';
-}
