--- lib/Symfony/ResourceConstants.php
@@ -1,8 +1,0 @@
-<?php
-namespace Ridibooks\OAuth2Resource\Symfony;
-
-class ResourceConstants
-{
-    const ACCESS_TOKEN_INFO_KEY = 'access_token_info';
-    const ACCESS_TOKEN_KEY = 'ridi-at';
-}
